pyproject.toml
[project]
name = "conversationalstack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-openai>=0.3.25",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.0",
]

[tool.uv.workspace]
members = [
    "frontend",
    "backend",
]
